Rejig models again

+++ db/migrate/20231021082833_create_regions.rb
@@ -2,7 +2,7 @@ class CreateRegions < ActiveRecord::Migration[6.1]
   def change
     create_table :regions do |t|
       t.string :name
-      t.integer :region_id, index: true
+      t.bigint :region_id, index: true
 
       t.timestamps
     end
