+++ db/migrate/20231021082838_create_areas.rb
@@ -2,7 +2,7 @@ class CreateAreas < ActiveRecord::Migration[6.1]
   def change
     create_table :areas do |t|
       t.string :name
-      t.integer :area_id, index: true
+      t.bigint :area_id, index: true
       t.references :region, null: false, foreign_key: true
 
       t.timestamps
